pyproject.toml
[tool.poetry]
name = "repl_python3_GunnHacks"
version = "0.1.0"
description = ""
authors = ["nicholasxwang <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
waitress = "^2.0.0"
wikipedia = "^1.4.0"
googlemaps = "^4.6.0"
poetry = "^1.1.12"
google-cloud-vision = "^2.6.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
